Refactor to run roll
