+++ app/src/main/java/ru/beregstudio/diceroller/MainActivity.kt
@@ -1,10 +1,12 @@
 package ru.beregstudio.diceroller
 
-import android.media.MediaPlayer
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
-import android.widget.*
+import android.widget.Button
+import android.widget.LinearLayout
+import android.widget.SeekBar
+import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 
 const val SIX_SIDES: Int = 6
@@ -19,12 +21,12 @@ class MainActivity : AppCompatActivity() {
         val rollNumber: TextView = findViewById(R.id.diceNumber)
         val rollButton: Button = findViewById(R.id.diceButton)
         val seekDice: SeekBar = findViewById(R.id.seekBar)
-
-        roller(DEFAULT_NUMBER_OF_DICE, layout, rollNumber)
+        val roll = Roll(layout, rollNumber, this)
+        roll.roller(DEFAULT_NUMBER_OF_DICE)
 
         rollButton.setOnClickListener {
             layout.removeAllViews()
-            roller(seekDice.progress + 1, layout, rollNumber)
+            roll.roller(seekDice.progress + 1)
         }
         seekDice.setOnSeekBarChangeListener(/* l = */ object : SeekBar.OnSeekBarChangeListener {
             /**
@@ -35,8 +37,9 @@ class MainActivity : AppCompatActivity() {
              */
             override fun onProgressChanged(seekBar: SeekBar, i: Int, b: Boolean) {
                 layout.removeAllViews()
-                roller(seekBar.progress + 1, layout, rollNumber)
+                roll.roller(seekDice.progress + 1)
             }
+
             /**
              * Событие воспроизведение трека при старте
              *
@@ -48,6 +51,7 @@ class MainActivity : AppCompatActivity() {
                  * Событие не применяется
                  */
             }
+
             /**
              * Событие при остановки воспроизведения
              *
@@ -74,58 +78,6 @@ class MainActivity : AppCompatActivity() {
             else -> 2
         }
         return super.onOptionsItemSelected(item)
-    }
-
-    /**
-     * Отрисовка кубика на экране
-     *
-     * Функция принимает атрибуты количество кубиков [numOfDice], лейаут [layout] и подпись [rollNumber]
-     * @param layout
-     * @param numOfDice
-     * @param rollNumber
-     */
-    private fun roller(numOfDice: Int, layout: LinearLayout, rollNumber: TextView) {
-        val diceRun = arrayListOf<Dice>()
-        var sum = 0
-        for (i in 1..numOfDice) {
-            diceRun.add(getDice())
-        }
-        println("Start debug")
-        for (dice in diceRun) {
-            layout.addView(dice.image)
-            rollNumber.text = dice.diceRoll.toString()
-            sum += dice.diceRoll
-            rollNumber.text = sum.toString()
-            println(dice.diceRoll.toString())
-            println("Image size:")
-            dice.setDiceSize()
-            println(dice.image.layoutParams.width.toString())
-            println("Image size end")
-        }
-        getDiceRollSound().start()
-        diceRun.clear()
-        println("Stop debug")
-
-    }
-
-    /**
-     * Получение нового кубика
-     *
-     * Функция возвращает объект [Dice] c присвоенным колличеством сторон.
-     * @return [Dice]
-     */
-    private fun getDice(): Dice {
-        return Dice(SIX_SIDES, diceset, this)
-    }
-
-    /**
-     * Проигрывание звука броска кубика
-     *
-     * Функция возвращает объект [MediaPlayer] c заданным треком.
-     * @return [MediaPlayer]
-     */
-    private fun getDiceRollSound(): MediaPlayer {
-        return MediaPlayer.create(this, R.raw.igralnaya_kost_upala)
     }
 
 }
